Fixing functions bit_floor_ul and bit_floor_ull in bit_manip_fuzzer_64bit_input

+++ fuzz/bit_manip/bit_manip_fuzzer_64bit_input.c
@@ -289,11 +289,11 @@ int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
     if (get_req_bit_width_ull_result != (ULLONG_WIDTH - (value == 0ULL ? ULLONG_WIDTH : count_leading_zeros_ull(value)))) __builtin_trap();
 
     // Fuzzing bit_floor_ul
-    uint32_t bit_floor_ul_result = bit_floor_ul(value);
+    uint64_t bit_floor_ul_result = bit_floor_ul(value);
     if (bit_floor_ul_result != (value == 0UL ? 0UL : 1UL << (ULONG_WIDTH - 1UL - count_leading_zeros_ul(value)))) __builtin_trap();
 
     // Fuzzing bit_floor_ull
-    uint32_t bit_floor_ull_result = bit_floor_ull(value);
+    uint64_t bit_floor_ull_result = bit_floor_ull(value);
     if (bit_floor_ull_result != (value == 0ULL ? 0ULL : 1ULL << (ULLONG_WIDTH - 1ULL - count_leading_zeros_ull(value)))) __builtin_trap();
 
     return 0;
